[Refactor] GraduateImage path
