+++ src/pages/graduates/components/GraduateImage/index.tsx
